refactor: remove unnecessary files
