--- apps/docs/src/components/HeadTags.tsx
@@ -1,55 +1,0 @@
-const HeadTags = () => {
-  return (
-    <>
-      <link
-        rel="apple-touch-icon"
-        sizes="180x180"
-        href="/favicon/apple-touch-icon.png"
-      />
-      <link
-        rel="icon"
-        type="image/png"
-        sizes="32x32"
-        href="/favicon/favicon-32x32.png"
-      />
-      <link
-        rel="icon"
-        type="image/png"
-        sizes="16x16"
-        href="/favicon/favicon-16x16.png"
-      />
-      <link rel="manifest" href="/favicon/site.webmanifest" />
-      <link
-        rel="mask-icon"
-        href="/favicon/safari-pinned-tab.svg"
-        color="#1B1F23"
-      />
-      <meta name="msapplication-TileColor" content="#1B1F23" />
-      <meta name="theme-color" content="#ffffff" />
-      <meta
-        property="og:image"
-        content="https://outstatic.com/images/og-image.png"
-      />
-      <meta
-        property="og:image:alt"
-        content="Outstatic - a static CMS for Next.js"
-      />
-      <meta name="twitter:card" content="summary_large_image" />
-      <meta name="twitter:site" content="@outstatic" />
-      <meta
-        name="twitter:description"
-        content="Write. Publish. It's live! Get a complete dashboard to manage your Next.js static website."
-      />
-      <meta
-        name="twitter:title"
-        content="Outstatic - A Static Site CMS for Next.js."
-      />
-      <meta
-        name="twitter:image"
-        content="https://outstatic.com/images/og-image.png"
-      />
-    </>
-  )
-}
-
-export default HeadTags
